1. Eigen factor not finished, may finish after. 2. fix the bug in derivative: mix the order of rotation and translation. 3. finish ceres factor, tested. 4. not sure if imu factor make influence, optimization and prediction must be added at the same time. 5. seems sliding window takes the mail role

+++ include/EventCamera.h
@@ -45,7 +45,7 @@ namespace CannyEVIT
         cv::Mat distortion_parameter_;
         DistortionModel distortion_type_;
         cv::Mat undistortion_map1_, undistortion_map2_;
-        Eigen::MatrixXd undistort_recitify_mask_;
+        Eigen::MatrixXi undistort_recitify_mask_;
 
         Eigen::Matrix4d Tbc_;
         Eigen::Matrix4d Tcb_;
@@ -58,7 +58,7 @@ namespace CannyEVIT
             return P_;
         }
 
-        inline Eigen::MatrixXd& getUndistortRectifyMask(){
+        inline Eigen::MatrixXi& getUndistortRectifyMask(){
             return undistort_recitify_mask_;
         }
